requirements.txt
# Django Core
Django>=5.0,<6.0
django-environ>=0.11.2

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
drf-spectacular>=0.27.0  # API Documentation

# CORS
django-cors-headers>=4.3.1

# File Storage
django-cloudinary-storage>=0.3.0
cloudinary>=1.36.0

# Static Files
whitenoise>=6.6.0

# CAPTCHA Protection
django-turnstile>=0.2.1  # Cloudflare Turnstile

# HTML Sanitization
bleach>=6.0.0  # XSS protection

# IP Address Handling
django-ipware>=7.0.1  # Accurate client IP detection (防止IP欺骗)

# Async Tasks & Caching
celery[redis]>=5.3.0  # Async task queue with Redis support
django-redis>=5.4.0  # Redis cache backend

# Social Authentication
social-auth-app-django>=5.4.0  # Google OAuth integration

# WebSockets & Real-time
channels>=4.0.0  # Django Channels for WebSockets
channels-redis>=4.1.0  # Redis channel layer
daphne>=4.0.0  # ASGI server for Channels

# Utilities
requests>=2.31.0  # For API calls
python-dateutil>=2.8.2
pillow>=10.1.0  # Image processing
pytz>=2023.3

# Security (Optional)
# django-ratelimit>=4.1.0  # Rate limiting
# django-defender>=0.9.7  # Login protection

# Development (Optional)
# django-debug-toolbar>=4.2.0
